+++ tests/unit/test_credit_card.py
@@ -12,7 +12,7 @@ class TestCreditCard(unittest.TestCase):
 
     def test_create_signature(self):
         expected = ["billing_address_id", "cardholder_name", "cvv", "expiration_date", "expiration_month",
-            "expiration_year", "number", "token", "venmo_sdk_payment_method_code",
+            "expiration_year", "number", "token", "venmo_sdk_payment_method_code",  # NEXT_MJOR_VERSION remove venmo_sdk_payment_method_code
             "device_data", "payment_method_nonce",
             "device_session_id", "fraud_merchant_id",
             {
@@ -26,7 +26,7 @@ class TestCreditCard(unittest.TestCase):
                 "fail_on_duplicate_payment_method",
                 "make_default",
                 "skip_advanced_fraud_checking",
-                "venmo_sdk_session",
+                "venmo_sdk_session", # NEXT_MJOR_VERSION remove venmo_sdk_session
                 "verification_account_type",
                 "verification_amount",
                 "verification_merchant_account_id",
@@ -44,7 +44,7 @@ class TestCreditCard(unittest.TestCase):
 
     def test_update_signature(self):
         expected = ["billing_address_id", "cardholder_name", "cvv", "expiration_date", "expiration_month",
-            "expiration_year", "number", "token", "venmo_sdk_payment_method_code",
+            "expiration_year", "number", "token", "venmo_sdk_payment_method_code",  # NEXT_MJOR_VERSION remove venmo_sdk_payment_method_code
             "device_data", "payment_method_nonce",
             "device_session_id", "fraud_merchant_id",
             {
@@ -58,7 +58,7 @@ class TestCreditCard(unittest.TestCase):
                 "fail_on_duplicate_payment_method",
                 "make_default",
                 "skip_advanced_fraud_checking",
-                "venmo_sdk_session",
+                "venmo_sdk_session", # NEXT_MJOR_VERSION remove venmo_sdk_session
                 "verification_account_type",
                 "verification_amount",
                 "verification_merchant_account_id",